Add support for SA:MP 0.3.7-R2 and improve type safety
- Added SAMPVersionIndex entry and offsets for 0.3.7-R2
- Updated GetSAMPVersionIndex and IsSupportedSAMPVersion for R2
- Changed GetPlayerId to return uint16_t and 0xFFFF on failure
- Updated header declaration for GetPlayerId
- Updated te_sdk_rel.lib to match code changes

+++ #TE SDK/te-samp.h
@@ -16,8 +16,8 @@ namespace te::sdk::helper::samp
 	// Get local player name
 	const char* GetPlayerName();
 
-	// Get local player ID (-1 on failure)
-	int GetPlayerId();
+	// Get local player ID as its true uint16 width (0xFFFF = INVALID_PLAYER_ID on failure)
+	uint16_t GetPlayerId();
 
 	// Check if connected to a server (game state >= 5)
 	bool IsGameLoaded();
